+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.11.15 (unreleased)
+
+- Remove invalid `unreachable!` call causing panic in some combination or for loop and specific filters
+- Fix macros loading in parent templates and using them in child ones
+- Fix macros loading other macros not working when called in inheritance
+- Mark `Context::add` as deprecated and do not display it in the docs anymore (aka TIL the `deprecated` attribute)
+- Fix `__tera_context` not getting all the available context (`set`, `forloop` etc)
+- Better error message when variable indexing fails
+
 ## 0.11.14 (2018-09-02)
 
 - Remove stray println